tracker caller is now a function, removed redundant variables and added function description
tracker caller is now a function with the filestub, nframes and
set_length as variables. Other variables are made optional with standard
values.
I removed the redundant variables widthcut, a1, pos and redundant links
to param for the ms error calculator noise_sz and widthcut. I updated
the function descriptions to the Summary, Input, Output form. It still
has to be revised completely for tracker caller and mserror calculator