InputParser and Variable Names
The optinal arguments in the function call are now governed by
inputParser. Therefore it is easier to add optional arguments in the
function call to add one for PlotOpt for example you should enter:
tracker_caller_4QM(filestub, nframes,'PlotOpt','none'). The variables in
the functioncall are changed to CamelCase to improve readability.