Corrected shifted data.
In shifting the particle there was a mistake in the coordinates put in
for interpolation. These coordinates were calculated by adding the shift
to a coordinate system.
However, a minus is needed here, because shifting the particle to the
right
with interp2, means that the coordinates should be shifted to the left.
Its like taking a picture. If you want an object to shift to the right
of your picture, you should move your camera to the left.