Boxsize and Deltafit
For accuracy the boxsize of the 4QM method should be bigger. To make it
easier to adjust this, we now define the standard box around 4QM to be
the feature size. DeltaFit can be given a positive value to increase
edge around the particle. Setting it to 9 would increase all sides with
9 pixels.